// circabc-frontend/src/app/core/generated/circabc/model/searchConfig.ts
/**
 * CIRCABC REST API
 * This is the first version of the CIRCABC REST API used by the new User Interface 
 *
 * The version of the OpenAPI document: 4.2.4.3
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { KeywordDefinition } from './keywordDefinition';


export interface SearchConfig { 
    searchName?: string;
    searchFor?: string;
    searchIn?: SearchConfig.SearchInEnum;
    language?: string;
    creationDateFrom?: string;
    creationDateTo?: string;
    lastModificationFrom?: string;
    lastModificationTo?: string;
    creatorUser?: User;
    keywords?: Array<KeywordDefinition>;
    status?: SearchConfig.StatusEnum;
}
export namespace SearchConfig {
    export type SearchInEnum = 'All' | 'library' | 'forums' | 'infoService' | 'agenda';
    export const SearchInEnum = {
        All: 'All' as SearchInEnum,
        Library: 'library' as SearchInEnum,
        Forums: 'forums' as SearchInEnum,
        InfoService: 'infoService' as SearchInEnum,
        Agenda: 'agenda' as SearchInEnum
    };
    export type StatusEnum = 'All' | 'library' | 'forums' | 'infoService' | 'agenda';
    export const StatusEnum = {
        All: 'All' as StatusEnum,
        Library: 'library' as StatusEnum,
        Forums: 'forums' as StatusEnum,
        InfoService: 'infoService' as StatusEnum,
        Agenda: 'agenda' as StatusEnum
    };
}